design: change to compiled library
