+++ doc/src/examples/CMakeLists.txt
@@ -1,10 +1,10 @@
 add_executable(simple_example simple.cpp)
 target_include_directories(simple_example PRIVATE ${CMAKE_SOURCE_DIR}/include)
-target_link_libraries(simple_example PRIVATE Threads::Threads)
+target_link_libraries(simple_example PRIVATE Threads::Threads ${LIB})
 
 add_executable(hooks_example hooks.cpp)
 target_include_directories(hooks_example PRIVATE ${CMAKE_SOURCE_DIR}/include)
-target_link_libraries(hooks_example PRIVATE Threads::Threads)
+target_link_libraries(hooks_example PRIVATE Threads::Threads ${LIB})
 
 add_custom_target(examples
   COMMAND
