+++ tests/CMakeLists.txt
@@ -10,10 +10,10 @@ add_executable(${TESTS_EXEC}
 )
 
 target_link_libraries(${TESTS_EXEC} PRIVATE
+  ${LIB}
   gtest_main
   Threads::Threads
 )
-target_include_directories(${TESTS_EXEC} PRIVATE ${CMAKE_SOURCE_DIR}/include
-  ${CMAKE_SOURCE_DIR}/tests)
+target_include_directories(${TESTS_EXEC} PRIVATE ${CMAKE_SOURCE_DIR}/tests)
 
 add_custom_target(check ${TESTS_EXEC} DEPENDS ${TESTS_EXEC} COMMENT "Running tests...")
